+++ resources/views/admin/sizecartinfo/update_sizecartinfo.blade.php
@@ -42,12 +42,13 @@
           @foreach ($sizecarts as $sizecart)
               
           <tr>
-            <th scope="row">{{$sizecart->id}}</th>
+            {{-- <th scope="row">{{$sizecart->id}}</th> --}}
+            <th scope="row">{{$loop->iteration}}</th>
             <td>{{$sizecart->size}}</td>
             <td>{{$sizecart->foot_length}}</td>
             <td>
               <a href="/admin/sizecartinfo/update_sizecart/{{$sizecart->id}}" class="btn btn-warning">update</a>
-              <a href="/admin/sizecartinfo/delete_sizecart/{{$sizecart->id}}" class="btn btn-info">delete</a>
+              <a onclick="return confirm('Are you sure?')" href="/admin/sizecartinfo/delete_sizecart/{{$sizecart->id}}" class="btn btn-info">delete</a>
             </td>
           </tr>
           @endforeach
